fix for issue 1 and experimental support for project update/delete
Not sure that the API supports updating or deleting the project itself, but, if it does, these changes should work.
Also added method to add a story to a project as a convenience.
Lastly, addresses issue #1 where the iteration was not correctly deserializing the list of stories for an iteration.