major changes and enhancements (UNTESTED)
NOTE: This is in a separate branch, as it is almost completely untested.
Includes attachment upload, notes (getting), and data migration functions.
If everything works, you can copy a story from one project to another, including
the tasks, notes, and attachments.  But it may not work...

The attachments is completely untested, and this uses some features of
the api that I am assuming return data but are undocumented (getting notes mainly).

Also includes a cloning function for all models to allow in-memory
serialization/deserialization back to a new object instance.

The only breaking changes should be for notes, I believe.